Update seg_hrnet.py to change np.int to np.int64

+++ models/seg_hrnet.py
@@ -393,7 +393,7 @@ class HighResolutionNet(nn.Module):
         self.stage4, pre_stage_channels = self._make_stage(
             self.stage4_cfg, num_channels, multi_scale_output=True)
 
-        last_inp_channels = np.int(np.sum(pre_stage_channels))
+        last_inp_channels = np.int64(np.sum(pre_stage_channels))
 
     ## GX: one dimension matrix converts pre to pos.
     ## GX: if channel numbers are equal, pass it directly.
